requirements.txt
aiocontextvars==0.2.2
azure-common==1.1.23
azure-core==1.9.0
azure-eventhub==5.0.1
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.6.0
azure-storage-common==1.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
contextvars==2.4
cryptography==3.0
elasticsearch==7.8.0
idna==2.10
immutables==0.14
isodate==0.6.0
loguru==0.5.1
msrest==0.6.19
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.3.0.36
protobuf==3.14.0
psycopg2==2.8.5
pycparser==2.20
pymongo==3.11.2
python-dateutil==2.8.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
uamqp==1.2.9
urllib3==1.25.10
